Use more strict compiler warnings

+++ src/xvc_dec_lib/syntax_reader.cc
@@ -281,7 +281,7 @@ void SyntaxReader::ReadCoeffSubblock(const CodingUnit &cu, YuvComponent comp,
             static_cast<Coeff>(ReadCoeffRemainExpGolomb(golomb_rice_k));
           if (subblock_coeff[i] > 3 * (1 << golomb_rice_k) &&
               !Restrictions::Get().disable_transform_adaptive_exp_golomb) {
-            golomb_rice_k = std::min(golomb_rice_k + 1, (uint32_t)4);
+            golomb_rice_k = std::min(golomb_rice_k + 1, 4u);
           }
         }
         if (subblock_coeff[i] >= 2) {
